+++ src/nvim/hardcopy.c
@@ -32,7 +32,7 @@
 #include "nvim/memline.h"
 #include "nvim/memory.h"
 #include "nvim/message.h"
-#include "nvim/misc1.h"
+#include "nvim/os/os.h"
 #include "nvim/misc2.h"
 #include "nvim/garray.h"
 #include "nvim/option.h"
