+++ src/nvim/mark.c
@@ -30,7 +30,6 @@
 #include "nvim/memline.h"
 #include "nvim/memory.h"
 #include "nvim/message.h"
-#include "nvim/misc1.h"
 #include "nvim/misc2.h"
 #include "nvim/normal.h"
 #include "nvim/option.h"
