+++ src/nvim/memfile.c
@@ -56,7 +56,6 @@
 #include "nvim/fileio.h"
 #include "nvim/memline.h"
 #include "nvim/message.h"
-#include "nvim/misc1.h"
 #include "nvim/misc2.h"
 #include "nvim/memory.h"
 #include "nvim/os_unix.h"
